Refatoração de codigo e remoção de classes não utilizadas.
