+++ Agenda/app/src/main/java/br/com/alura/agenda/database/converter/ConversorCalendar.java
@@ -8,7 +8,10 @@ public class ConversorCalendar {
 
     @TypeConverter
     public Long paraLong(Calendar valor){
-        return valor.getTimeInMillis();
+        if(valor != null){
+            return valor.getTimeInMillis();
+        }
+        return null;
     }
 
     @TypeConverter
