+++ Agenda/app/src/main/java/br/com/alura/agenda/model/Aluno.java
@@ -18,17 +18,6 @@ public class Aluno implements Serializable {
     private String email;
     private Calendar momentoDeCadastro = Calendar.getInstance();
 
-    @Ignore
-    public Aluno(String nome, String telefone, String email){
-        this.nome = nome;
-        this.telefone = telefone;
-        this.email = email;
-    }
-
-    public Aluno() {
-
-    }
-
     public Calendar getMomentoDeCadastro() {
         return momentoDeCadastro;
     }
@@ -77,14 +66,5 @@ public class Aluno implements Serializable {
 
     public boolean temIdValido() {
         return id > 0;
-    }
-
-    public String getNomeCompleto() {
-        return nome;
-    }
-
-    public String dataFormatada(){
-        SimpleDateFormat formatador = new SimpleDateFormat("dd/MM/yyyy");
-        return formatador.format(momentoDeCadastro.getTime());
     }
 }
