+++ Agenda/app/src/main/java/br/com/alura/agenda/ui/adapter/ListaAlunosAdapter.java
@@ -47,7 +47,7 @@ public class ListaAlunosAdapter extends BaseAdapter {
 
     private void vincula(View view, Aluno aluno) {
         TextView nome = view.findViewById(R.id.item_aluno_nome);
-        nome.setText(aluno.getNomeCompleto() + " " + aluno.dataFormatada());
+        nome.setText(aluno.getNome());
         TextView telefone = view.findViewById(R.id.item_aluno_telefone);
         telefone.setText(aluno.getTelefone());
     }
